Finished README.md and formatting changes for service and repo layers

+++ DGTournamentTracker/src/main/java/com/skilldistillery/dgtournament/services/TournamentService.java
@@ -32,9 +32,9 @@ public interface TournamentService {
 
 	int countTop10Finishes();
 	
-	int countTotalPointsForAllTournaments();
+	double countTotalPointsForAllTournaments();
 	
-	int countTotalPointsPerYear(int year);
+	double countTotalPointsPerYear(int year);
 	
 	double calculateAveragePointsPerTournament();
